update Object to support alive and blocking player
This required some refactoring of the `move_by` method on Object
to avoid borrowing conflicts.

`move_by` has been moved out of Object as a plain function.

`handle_keys` was updated to accept Map for `move_by`, and all objects instead of only player

finally, `place_objects` now accepts a `Map` and checks if a spot is blocked on the map before creating a monster